+++ CoordFiles/coord.h
@@ -8,10 +8,10 @@
 
 class coord {
 public:
-    float x;
-    float y;
-    float z;
-    float w;
+    double x;
+    double y;
+    double z;
+    double w;
 };
